如果area是null，拆包就会包空指针异常

+++ src/main/java/org/gyh/forestry/dto/resp/ForestFire.java
@@ -12,9 +12,9 @@ import java.time.LocalDateTime;
 public class ForestFire {
     private String name;
     @Schema(description = "面积")
-    private double area;
+    private Float area;
     @Schema(description = "含水率")
-    private double moistureContent;
+    private Float moistureContent;
     /**
      * 温度
      */
